Add panic recovery test
